server: move sqlServer to new file
This is pure code movement. Next step will be untangling
`(*Server).Start` to contain an explicit call to `(*sqlServer).start()`
(a method which does not exist today).

I'm intentionally not documenting the various inputs to `sqlServer` yet
in anticipation of larger changes to them.

Release note: None
